+++ histogram_test.go
@@ -73,8 +73,8 @@ func TestHistogramValueMarshaling(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	exp := strings.Replace(strings.Replace(strings.Replace(histogramTestData,
-		" ", "", -1), "\n", "", -1), "\t", "", -1) + "\n"
+	exp := strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(
+		histogramTestData, " ", ""), "\n", ""), "\t", "") + "\n"
 	if buf.String() != exp {
 		t.Errorf("Expected JSON: %v, got: %v", exp, buf.String())
 	}